Add ollama ChatRequest and ChatResponse
This patch improves the template usage to provide a profile to the
ocstack ChatRequest.
It is possible to setup a profile, in the form of a golang template, and
pass it to set a context in the llm module call. By doing this we can
influence the reasoning based on some external input.
At the same time this patch provides some improvements and modules
refactoring.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>