Externalize local tools schema definition
This patch externalizes the way we define tool's schema. In particular
it is now possible to define an external tools.json file that contains
the definition and is imported when the tool is executed. It is
supported to merge multiple json tools definition into a single
map[string]any, and the implementation, as well as the Provider code
must be updated to actually call the tools implemented in utils.go.

Signed-off-by: Francesco Pantano <fpantano@redhat.com>