Add LLamaCPP provider
This patch enables the Ramalama usage by adding the LLAMACPP provider.
Ramalama uses llamacpp webserver to expose the LLM [1] and we can use
this information to build an appropriate provider. This also implements
a more generic interface.
To make this happen, I had to modify the existing provider.go to better
handling the history, and the refactory is not yet complete.
This patch also improves a few things:
1. provide a more compact default.tmpl agent profile
2. provide a more detailed agent.tmpl profile
3. DEBUG now prints information about history

[1] https://github.com/containers/ramalama/issues/265

Signed-off-by: Francesco Pantano <fpantano@redhat.com>